Peek at last arg to see if it's a Hash
This change preserves the existing behavior, just to make sure that the
tests still pass even though I'm peeking one level at last argument.

This sort of a peek is similar to what the Ruby VM does when it
commpiles a Ruby program to bytecode. For example:

    ❯ ruby --dump=insns 'foo({x: 12}, {y: 34})'
    == disasm: #<ISeq:<main>@foo.rb:1 (1,0)-(1,19)> (catch: FALSE)
    0000 putself                                                          (   1)[Li]
    0001 duphash                      {:x=>12}
    0003 putobject                    34
    0005 opt_send_without_block       <callinfo!mid:foo, argc:2, kw:[y], FCALL|KWARG>, <callcache>
    0008 leave

The Ruby VM records the second argument not as a value on the VM stack,
but with a special operand to the `opt_send_without_block` instruction
that remembers which explicit keywords were mentioned.

Which is to say: if the last argument to a method call is a Hash
literal, it already gets special treatment by the VM, so it's fine for
us to do the same.